# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Order Transaction API",
        "version": "1.0"
    },
    "host": "",
    "basePath": "/api/v1",
    "paths": {
...
        "/order": {
            "get": {
                "tags": [
                    "/order"
                ],
                "summary": "List Order",
                "description": "List Order",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "order_number",
                        "in": "query",
                        "description": "User ID cart owner",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Product exist on cart",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "description": "Product exist on cart",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Product exist on cart",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Product exist on cart",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "tracking_no",
                        "in": "query",
                        "description": "Product exist on cart",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "coupon_code",
                        "in": "query",
                        "description": "Product exist on cart",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 10
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                }
            },
            "post": {
                "tags": [
                    "/order"
                ],
                "summary": "Create new Order",
                "description": "Create new Order",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Coupon Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Coupon Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "phone_number",
                        "in": "formData",
                        "description": "Coupon Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "address",
                        "in": "formData",
                        "description": "Coupon Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "shipping_method",
                        "in": "formData",
                        "description": "Coupon Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "coupon_code",
                        "in": "formData",
                        "description": "Coupon Name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
...
        "/payment": {
            "get": {
                "tags": [
                    "/payment"
                ],
                "summary": "List Payment",
                "description": "List Payment",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "payment_date",
                        "in": "query",
                        "description": "Date of payment",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "Order User ID",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user_account",
                        "in": "query",
                        "description": "Sender Bank User Account",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user_bank_account",
                        "in": "query",
                        "description": "Sender Bank Name",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "order_id",
                        "in": "query",
                        "description": "Order ID",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 10
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Limit data",
                        "required": true,
                        "type": "number",
                        "default": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                }
            },
            "post": {
                "tags": [
                    "/payment"
                ],
                "summary": "Create new Payment",
                "description": "Create new Payment",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "formData",
                        "description": "Order ID",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "user_account",
                        "in": "formData",
                        "description": "Sender Bank User Account",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_bank_account",
                        "in": "formData",
                        "description": "Sender Bank Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "reference_no",
                        "in": "formData",
                        "description": "Reference No",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "notes",
                        "in": "formData",
                        "description": "Notes",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/payment/{id}/confirm": {
            "get": {
                "tags": [
                    "/order"
                ],
                "summary": "Confirm Payment",
                "description": "Confirm Payment",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/payment/{id}/reject": {
            "get": {
                "tags": [
                    "/payment"
                ],
                "summary": "Reject Payment",
                "description": "Reject Payment",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "payment ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/payment/{id}/cancel": {
            "get": {
                "tags": [
                    "/payment"
                ],
                "summary": "Cancel Payment",
                "description": "Cancel Payment",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
        "/payment/{id}": {
            "get": {
                "tags": [
                    "/payment"
                ],
                "summary": "View Payment detail",
                "description": "View Payment detail",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payment ID",
                        "required": true,
                        "type": "number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation success"
                    }
                },
                "security": [
                    {
                        "AccessToken": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "x-access-token",
            "in": "header"
        }
    },
...
}